restructure file layout to shrink file sizes
