+++ OneSila/core/schema/multi_tenant/mutations/__init__.py
@@ -1,0 +1,1 @@
+from .mutation_type import MultiTenantMutation
