+++ OneSila/core/schema/multi_tenant/mutations/mutation_classes.py
@@ -1,5 +1,5 @@
 from strawberry import UNSET
-from strawberry_django import auth as strawberry_auth
+
 from strawberry_django.resolvers import django_resolver
 from strawberry_django.mutations import resolvers
 from strawberry_django.auth.utils import get_current_user
@@ -8,30 +8,19 @@ from strawberry_django.utils.requests import get_request
 from strawberry_django.auth.exceptions import IncorrectUsernamePasswordError
 
 from django.db import transaction
-from django.core.exceptions import ValidationError
 from django.contrib import auth
 from django.contrib.auth.password_validation import validate_password
-from django.db.utils import IntegrityError
 from django.conf import settings
 
-from typing import cast, Type
-from asgiref.sync import async_to_sync, sync_to_async
+from asgiref.sync import async_to_sync
 
 from channels import auth as channels_auth
-from channels.db import database_sync_to_async
 
 from core.schema.core.mutations import create, type, DjangoUpdateMutation, \
     DjangoCreateMutation, GetMultiTenantCompanyMixin, default_extensions, \
     update, Info, models, Iterable, Any, IsAuthenticated
-from core.schema.core.mixins import GetQuerysetMultiTenantMixin
 from core.factories.multi_tenant import InviteUserFactory, RegisterUserFactory, \
     AcceptUserInviteFactory, EnableUserFactory, DisableUserFactory
-
-from .types.types import MultiTenantUserType, MultiTenantCompanyType
-from .types.input import MultiTenantUserInput, MultiTenantUserPartialInput, \
-    MultiTenantCompanyPartialInput, MultiTenantCompanyInput, \
-    MultiTenantInviteUserInput, MultiTenantCompanyMyInput, \
-    MultiTenantAcceptInviteInput, MultiTenantUserStatusInput
 
 
 class CleanupDataMixin:
@@ -179,61 +168,3 @@ class EnableUserMutation(DjangoUpdateMutation):
             fac = EnableUserFactory(user=instance)
             fac.run()
             return fac.user
-
-
-def register_my_multi_tenant_company():
-    extensions = [IsAuthenticated()]
-    return MyMultiTenantCompanyCreateMutation(MultiTenantCompanyMyInput, extensions=extensions)
-
-
-def update_my_multi_tenant_company():
-    extensions = default_extensions
-    return MyMultiTentantCompanyUpdateMutation(MultiTenantCompanyPartialInput, extensions=extensions)
-
-
-def update_me():
-    extensions = default_extensions
-    return UpdateMeMutation(MultiTenantUserPartialInput, extensions=extensions)
-
-
-def register_user():
-    extensions = []
-    return RegisterUserMutation(MultiTenantUserInput, extensions=extensions)
-
-
-def invite_user():
-    extensions = default_extensions
-    return InviteUserMutation(MultiTenantInviteUserInput, extensions=extensions)
-
-
-def accept_user_invitation():
-    extensions = default_extensions
-    return AcceptInvitationMutation(MultiTenantAcceptInviteInput, extensions=extensions)
-
-
-def disable_user():
-    extensions = default_extensions
-    return EnableUserMutation(MultiTenantUserStatusInput, extensions=extensions)
-
-
-def enable_user():
-    extensions = default_extensions
-    return DisableUserMutation(MultiTenantUserStatusInput, extensions=extensions)
-
-
-@type(name="Mutation")
-class MultiTenantMutation:
-    login: MultiTenantUserType = strawberry_auth.login()
-    logout = strawberry_auth.logout()
-
-    register_user: MultiTenantUserType = register_user()
-    register_my_multi_tenant_company: MultiTenantCompanyType = register_my_multi_tenant_company()
-
-    update_me: MultiTenantUserType = update_me()
-    update_my_multi_tenant_company: MultiTenantCompanyType = update_my_multi_tenant_company()
-
-    invite_user: MultiTenantUserType = invite_user()
-    accept_user_invitation: MultiTenantUserType = accept_user_invitation()
-
-    disable_user: MultiTenantUserType = disable_user()
-    enable_user: MultiTenantUserType = enable_user()
