+++ OneSila/core/tests/tests_schemas/mutations.py
@@ -1,27 +1,41 @@
 
 REGISTER_USER_MUTATION = """
-    mutation registerUser($username: String!, $password: String!, $language: String!) {
-      registerUser(data: {username: $username, password: $password, language: $language}) {
-        username
-        password
-        invitationAccepted
-        isMultiTenantCompanyOwner
-      }
+  mutation registerUser($username: String!, $password: String!, $language: String!) {
+    registerUser(data: {username: $username, password: $password, language: $language}) {
+      username
+      password
+      invitationAccepted
+      isMultiTenantCompanyOwner
     }
+  }
 """
 
 
 LOGIN_MUTATION = """
-    mutation login($username: String!, $password: String!) {
-        login(username: $username, password: $password){
-            username
-        }
+  mutation login($username: String!, $password: String!) {
+    login(username: $username, password: $password){
+      username
     }
+  }
 """
 
 
 LOGOUT_MUTATION = """
-    mutation logout {
-        logout
+  mutation logout {
+      logout
+  }
+"""
+
+
+ME_QUERY = """
+  query me {
+    me {
+      username
+      password
+      language
+      multiTenantCompany{
+        id
+      }
     }
+  }
 """
